Add dispatch tables for all dispatchable handles
Now the function pointer structs and the functions that use them
are separate. Needs a renaming pass. Extensions keep their own dispatch
table, but the dispatchable handles include all extensions as well.

Also began adding bitfield support, but need to redo it to special case
GeometryInstanceFlagsKHR because of course there has to be special cases.