Start 'basic' vulkan bindings, remove cpp17 folder
This is the beginning of the basic C++ vulkan bindings. Current changes compared to
the base api is using enum class, creating typesafe flags, initializing struct members
to zero and sType to the stype value.

No changes are planned to function definitions, but it is desired to create a volk-like
dispatch loader.